requirements.txt
torch
torchvision
torchaudio
scikit-learn
imbalanced-learn
joblib
pandas
numpy
pywebview
pyserial